[#41944] cli: Add artifact read command
This allows for easy inspection of the contents of an artifact. Additionally,
the passthrough for unknown commands was removed for integrity reasons
(mender-artifact will not know how to handle our delta artifacts, which could
cause corruption). The passthrough will still be done, but only for individual
commands (like read here).